+++ flask_wtf/form.py
@@ -1,122 +1,93 @@
-# coding: utf-8
 import warnings
 
-from flask import current_app, request, session
+from flask import current_app, request
+from flask import session
 from jinja2 import Markup
-from werkzeug.datastructures import MultiDict
-from wtforms.ext.csrf.form import SecureForm
-from wtforms.validators import ValidationError
+from werkzeug.datastructures import CombinedMultiDict, ImmutableMultiDict
+from wtforms import Form
+from wtforms.meta import DefaultMeta
 from wtforms.widgets import HiddenInput
 
 from ._compat import FlaskWTFDeprecationWarning, string_types, text_type
-from .csrf import generate_csrf, validate_csrf
+from .csrf import _FlaskFormCSRF
 
 try:
     from .i18n import translations
 except ImportError:
     translations = None  # babel not installed
 
+
 SUBMIT_METHODS = set(('POST', 'PUT', 'PATCH', 'DELETE'))
+_Auto = object()
 
 
-class _Auto(object):
-    """Placeholder for unspecified variables that should be set to defaults.
+class FlaskForm(Form):
+    """Flask-specific subclass of WTForms :class:`~wtforms.form.Form`.
 
-    Used when None is a valid option and should not be replaced by a default.
-    """
-    pass
-
-
-class FlaskForm(SecureForm):
-    """Flask-specific subclass of WTForms :class:`~wtforms.ext.csrf.form.SecureForm` class.
-
-    If ``formdata`` is not specified, this will use :attr:`flask.request.form` and
-    :attr:`flask.request.files`.  Explicitly pass ``formdata=None`` to prevent this.
-
-    :param csrf_context: a session or dict-like object to use when making
-        CSRF tokens. Default: :data:`flask.session`.
-
-    :param secret_key: a secret key for building CSRF tokens. If this isn't
-        specified, the form will take the first of these
-        that is defined:
-
-        * SECRET_KEY attribute on this class
-        * WTF_CSRF_SECRET_KEY config of Flask app
-        * SECRET_KEY config of Flask app
-        * session secret key
-
-    :param csrf_enabled: whether to use CSRF protection. If False, all
-        csrf behavior is suppressed.
-        Default: WTF_CSRF_ENABLED config value
+    If ``formdata`` is not specified, this will use :attr:`flask.request.form`
+    and :attr:`flask.request.files`.  Explicitly pass ``formdata=None`` to
+    prevent this.
     """
 
-    SECRET_KEY = None
-    TIME_LIMIT = None
+    class Meta(DefaultMeta):
+        csrf_class = _FlaskFormCSRF
+        csrf_context = session  # not used, provided for custom csrf_class
 
-    def __init__(self, formdata=_Auto, obj=None, prefix='', csrf_context=None,
-                 secret_key=None, csrf_enabled=None, **kwargs):
+        def __init__(self):
+            config = current_app.config.get
+            self.csrf = config('WTF_CSRF_ENABLED', True)
+            self.csrf_secret = config(
+                'WTF_CSRF_SECRET_KEY', current_app.secret_key
+            )
+            self.csrf_field_name = config('WTF_CSRF_FIELD_NAME', 'csrf_token')
+            self.csrf_time_limit = config('WTF_CSRF_TIME_LIMIT', 3600)
 
-        if csrf_enabled is None:
-            csrf_enabled = current_app.config.get('WTF_CSRF_ENABLED', True)
+        def wrap_formdata(self, form, formdata):
+            if formdata is _Auto:
+                if _is_submitted():
+                    if request.files:
+                        return CombinedMultiDict((
+                            request.files, request.form
+                        ))
+                    elif request.form:
+                        return request.form
+                    elif request.get_json():
+                        return ImmutableMultiDict(request.get_json())
 
-        self.csrf_enabled = csrf_enabled
+                return None
 
-        if formdata is _Auto:
-            if self.is_submitted():
-                formdata = request.form
-                if request.files:
-                    formdata = formdata.copy()
-                    formdata.update(request.files)
-                elif request.get_json():
-                    formdata = MultiDict(request.get_json())
-            else:
-                formdata = None
+            return formdata
 
-        if self.csrf_enabled:
-            if csrf_context is None:
-                csrf_context = session
-            if secret_key is None:
-                # It wasn't passed in, check if the class has a SECRET_KEY
-                secret_key = getattr(self, "SECRET_KEY", None)
+        def get_translations(self, form):
+            if not current_app.config.get('WTF_I18N_ENABLED', True):
+                return None
 
-            self.SECRET_KEY = secret_key
-        else:
-            csrf_context = {}
-            self.SECRET_KEY = ''
-        super(FlaskForm, self).__init__(
-            formdata, obj, prefix,
-            csrf_context=csrf_context,
-            **kwargs
-        )
+            return translations
 
-    def generate_csrf_token(self, csrf_context=None):
-        if not self.csrf_enabled:
-            return None
+    def __init__(self, formdata=_Auto, **kwargs):
+        csrf_enabled = kwargs.pop('csrf_enabled', None)
 
-        return generate_csrf(secret_key=self.SECRET_KEY)
+        if csrf_enabled is not None:
+            warnings.warn(FlaskWTFDeprecationWarning(
+                '"csrf_enabled" is deprecated and will be removed in 1.0. '
+                'Set "meta.csrf" instead.'
+            ), stacklevel=3)
+            kwargs.setdefault('meta', {}).setdefault('csrf', csrf_enabled)
 
-    def validate_csrf_token(self, field):
-        if not self.csrf_enabled:
-            return True
-
-        if getattr(request, 'csrf_valid', False):
-            # this is validated by CsrfProtect
-            return True
-
-        if not self.validate_csrf_data(field.data):
-            raise ValidationError(field.gettext('CSRF token missing'))
-
-    def validate_csrf_data(self, data):
-        """Check if the given data is a valid CSRF token."""
-
-        return validate_csrf(data, secret_key=self.SECRET_KEY, time_limit=self.TIME_LIMIT)
+        super(FlaskForm, self).__init__(formdata=formdata, **kwargs)
 
     def is_submitted(self):
         """Consider the form submitted if there is an active request and
         the method is ``POST``, ``PUT``, ``PATCH``, or ``DELETE``.
         """
 
-        return request and request.method in SUBMIT_METHODS
+        return _is_submitted()
+
+    def validate_on_submit(self):
+        """Call :meth:`validate` only if the form is submitted.
+        This is a shortcut for ``form.is_submitted() and form.validate()``.
+        """
+        return self.is_submitted() and self.validate()
 
     def hidden_tag(self, *fields):
         """Render the form's hidden fields in one call.
@@ -149,26 +120,17 @@ class FlaskForm(SecureForm):
 
                 yield f
 
-        return Markup(u'\n'.join(text_type(f) for f in hidden_fields(fields or self)))
+        return Markup(
+            u'\n'.join(text_type(f) for f in hidden_fields(fields or self))
+        )
 
-    def validate_on_submit(self):
-        """Call :meth:`validate` only if the form is submitted.
-        This is a shortcut for ``form.is_submitted() and form.validate()``.
-        """
-        return self.is_submitted() and self.validate()
 
-    @property
-    def data(self):
-        d = super(FlaskForm, self).data
-        # https://github.com/lepture/flask-wtf/issues/208
-        if self.csrf_enabled:
-            d.pop('csrf_token', None)
-        return d
+def _is_submitted():
+    """Consider the form submitted if there is an active request and
+    the method is ``POST``, ``PUT``, ``PATCH``, or ``DELETE``.
+    """
 
-    def _get_translations(self):
-        if not current_app.config.get('WTF_I18N_ENABLED', True):
-            return None
-        return translations
+    return bool(request) and request.method in SUBMIT_METHODS
 
 
 class Form(FlaskForm):